backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
sentence-transformers
numpy
chromadb
redis
